+++ custom_components/soundtouchplus/const.py
@@ -16,13 +16,14 @@ DEFAULT_TIMEOUT = 15
 
 # custom service names.
 SERVICE_AUDIO_TONE_LEVELS = "audio_tone_levels"
+SERVICE_MUSICSERVICE_STATION_LIST = "musicservice_station_list"
 SERVICE_PLAY_CONTENTITEM = "play_contentitem"
 SERVICE_PLAY_HANDOFF = "play_handoff"
 SERVICE_PLAY_TTS = "play_tts"
 SERVICE_PLAY_URL = "play_url"
-SERVICE_PRESETLIST = "preset_list"
+SERVICE_PRESET_LIST = "preset_list"
 SERVICE_REBOOT_DEVICE = "reboot_device"
-SERVICE_RECENTLIST = "recent_list"
+SERVICE_RECENT_LIST = "recent_list"
 SERVICE_REMOTE_KEYPRESS = "remote_keypress"
 SERVICE_SNAPSHOT_RESTORE = "snapshot_restore"
 SERVICE_SNAPSHOT_STORE = "snapshot_store"